# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MobileCRM API Documentation",
        "description": "MobileCRM API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/business/hotel/rate-plans": {
            "get": {
                "tags": [
                    "Hotel - Rate Plans"
                ],
                "summary": "List Rate Plans",
                "operationId": "74d1a582e5816effc4215d4a789cee5a",
                "responses": {
                    "200": {
                        "description": "Rate plans retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Hotel - Rate Plans"
                ],
                "summary": "Create Rate Plan",
                "operationId": "7d3de70090e02ed4cbb995be5956f43c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "start_date",
                                    "end_date",
                                    "modifier_type",
                                    "modifier_value"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "modifier_type": {
                                        "type": "string",
                                        "enum": [
                                            "fixed",
                                            "percentage"
                                        ]
                                    },
                                    "modifier_value": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rate plan created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/business/hotel/rate-plans/{id}": {
            "get": {
                "tags": [
                    "Hotel - Rate Plans"
                ],
                "summary": "Get Rate Plan",
                "operationId": "4825516263b163d652a82342f69601b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rate plan retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Hotel - Rate Plans"
                ],
                "summary": "Update Rate Plan",
                "operationId": "ed7128e699db7e5218bbe9672990b378",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rate plan updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Hotel - Rate Plans"
                ],
                "summary": "Delete Rate Plan",
                "operationId": "c1b98e2c2cde724937f9fed6d3e91c6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rate plan deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer Token to authenticate",
                "scheme": "bearer"
            }
        }
    },
...
}